# widget.SimpleRenderer: use an array instead of a slice