### Description:
The widget.SimpleRenderer only renders one object and was using a slice to hold it (24 bytes).
By using a single item array instead, its size comes down to 16 bytes.

### Checklist:

- [X] Tests included.
- [X] Lint and formatter run with no errors.
- [X] Tests all pass.