@andydotxyz 2024-09-07
Thanks for this suggestion. However the call to `Objects()` happens on a very frequent basis, won't the CPU required constantly converting [1] to [] negate the benefits of saved ram?