@pierrec 2024-09-07
I dont htink so but there is only one way to find out... Will add a benchmark.