@pierrec 2024-09-07
Here are the benchmark results:

```
func BenchmarkSimpleRenderObjects(b *testing.B) {
	var obj fyne.CanvasObject
	r := widget.NewSimpleRenderer(obj)
	for i := 0; i < b.N; i++ {
		r.Objects()
	}
}
```

```
go test -bench Simple -count 10 > before.txt
goos: darwin
goarch: amd64
pkg: fyne.io/fyne/v2/internal/widget
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3875 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3872 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3871 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3875 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3874 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3876 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3876 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3877 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3872 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3873 ns/op
PASS
ok  	fyne.io/fyne/v2/internal/widget	6.401s
```

```
go test -bench Simple -count 10 > after.txt
goos: darwin
goarch: amd64
pkg: fyne.io/fyne/v2/internal/widget
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3884 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3874 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3872 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3872 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3871 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3879 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3885 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3891 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3892 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         0.3883 ns/op
PASS
ok  	fyne.io/fyne/v2/internal/widget	6.419s
```

```
~/go/bin/benchstat before.txt after.txt
goos: darwin
goarch: amd64
pkg: fyne.io/fyne/v2/internal/widget
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
                       │  before.txt  │           after.txt            │
                       │    sec/op    │    sec/op     vs base          │
SimpleRenderObjects-12   0.3875n ± 0%   0.3881n ± 0%  ~ (p=0.157 n=10)
```