@Jacalz 2024-09-07
Those benchmarks are not helping you anything. Because you are not saving the result from the call to Objects(), it is optimising away the entire call. That is why you are seeing abnormal low numbers that are almost as fast as the CPU can execute an instruction.