@pierrec 2024-09-07
> Those benchmarks are not helping you prove anything. Because you are not saving the result from the call to Objects(), it is optimising away the entire call. That is why you are seeing abnormal low numbers that are the same before and after as well as being as fast as the CPU can execute a single instruction.
> 
> You need to save the objects slice to a local variable on each iteration and save that variable to a global variable at the end of the benchmark if you want correct results.

Ah yes. No impact still after updating the benchmark.

```
var benchmarkObjects []fyne.CanvasObject

func BenchmarkSimpleRenderObjects(b *testing.B) {
	var obj fyne.CanvasObject
	r := widget.NewSimpleRenderer(obj)
	for i := 0; i < b.N; i++ {
		benchmarkObjects = r.Objects()
	}
}
```

```
cat before.txt
goos: darwin
goarch: amd64
pkg: fyne.io/fyne/v2/internal/widget
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.164 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.164 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
PASS
ok  	fyne.io/fyne/v2/internal/widget	14.959s
```

```
cat after.txt
goos: darwin
goarch: amd64
pkg: fyne.io/fyne/v2/internal/widget
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSimpleRenderObjects-12    	954153519	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.163 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.170 ns/op
BenchmarkSimpleRenderObjects-12    	1000000000	         1.162 ns/op
PASS
ok  	fyne.io/fyne/v2/internal/widget	14.898s
```

```
~/go/bin/benchstat before.txt after.txt
goos: darwin
goarch: amd64
pkg: fyne.io/fyne/v2/internal/widget
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
                       │ before.txt  │           after.txt           │
                       │   sec/op    │   sec/op     vs base          │
SimpleRenderObjects-12   1.163n ± 0%   1.162n ± 0%  ~ (p=0.496 n=10)
```