@andydotxyz 2024-09-07
Is this good now @Jacalz? I am no good with benchmarks :)