@Jacalz 2024-09-07
I mean, the benchmarks mostly look good but would have been more accurate if they used a local variable in the loop and then assigned that local variable to the global one at the end of the benchmark (right now it is including the overhead of global assignment in each benchmark) but it probably won't make much difference to the results. However, that is likely of no importance.

I'm a bit on the fence about this though. Feels a bit like code churn and extra complexity for the sake of very little to gain for no extra performance? You need an insane amount of renderers for these extra bytes to matter in any memory usage situation.