@Jacalz 2024-09-07
For the record: I have a vague recollection of trying to do this optimization sometime before and avoiding it due to being afraid that the conversion to a slice would allocate. It might be possible to gain more insight by looking at the generated assembly in something like https://go.godbolt.org/ or use the linker flag to have Go compiler tell us when allocations are happening (the latter can also be done interactively in VSCode quite nicely). That would likely be a better source of information than a small micro benchmark but I don't know.