@pierrec 2024-09-07
I am new to Fyne, so not sure how often the SimpleRenderer is actually used. :)
However this seems like it is not worth it (the benchmark does not show any allocation either (go1.22) but it might on other versions). Going to close this so as not make you waste more time. Thanks for the feedback though.