@Jacalz 2024-09-07
Thanks for looking at it and trying to optimize Fyne. Even if this PR wasn't merged, it is still greatly appreciated work. Feel free to look into more areas and ask any questions when the arise. Welcome to Fyne :)